requirements.txt
bcrypt==3.2.0
boto==2.49.0
boto3==1.17.99
botocore==1.20.99
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
Django==2.2
django-middleware-global-request==0.1.2
djangorestframework==3.12.2
et-xmlfile==1.0.1
html5lib==1.0b10
httplib2==0.18.1
idna==2.7
jdcal==1.4.1
jmespath==0.10.0
mysqlclient==2.0.1
numpy==1.20.1
openpyxl==3.0.5
pandas==1.2.2
paramiko==2.7.2
Pillow==8.1.0
pycparser==2.20
pyfcm==1.4.7
PyMySQL==1.0.2
PyNaCl==1.4.0
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2021.1
razorpay==1.2.0
reportlab==3.5.59
requests==2.19.1
s3transfer==0.4.2
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
webencodings==0.5.1
xhtml2pdf==0.2b1
xlrd==2.0.1
xmltodict==0.11.0